no error no ko no norm error but have crash problem

+++ real_42/libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: ukim <ukim@42seoul.kr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/04 00:34:26 by ukim              #+#    #+#             */
-/*   Updated: 2020/10/07 14:47:36 by ukim             ###   ########.fr       */
+/*   Updated: 2020/10/07 20:04:06 by ukim             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	unsigned char	s2_letter;
 
 	i = 0;
-	while (s1[i] == s2[i] && s1[i] != '\0' && s2[i] != '\0' && i < n)
+	while (s1[i] == s2[i] && s1[i] != '\0' && s2[i] != '\0' && (size_t)i < n)
 	{
 		i++;
 	}
